fix documentation naming
